Improve test script

+++ test_backup.sh
@@ -2,48 +2,62 @@
 
 set -e
 
+if [[ $# -ne 1 ]]; then
+    echo "usage: test_backup.sh <dir to backup>"
+    exit 1
+fi
+
+: ${TMPDIR:=/tmp}
+
 # unencrypted
-export BK_DIR=/tmp/bk
+SRC_DIR=$argv[1]
+RESTORE_DIR=$TMPDIR/bk_restore
+BK=(bk --verbose)
+
+export BK_DIR=$TMPDIR/bk
+
 rm -rf $BK_DIR
 mkdir $BK_DIR
-bk init
-bk backup  go ~/go/src 
-bk backup  go2 ~/go/src
-bk backup --base go2  go3 ~/go/src
-bk list  
-bk fsck  
 
-rm -rf /tmp/go 
-bk restore  go /tmp/go 
-(cd ~/go/src; for i in **/*(^/); do cmp $i /tmp/go/$i; done)
+$BK init
+$BK backup test $SRC_DIR
+sleep 1 # make sure times don't match
+$BK backup test $SRC_DIR
+sleep 1
+$BK backup --base test test $SRC_DIR
+$BK list  
+$BK fsck  
 
-rm -rf /tmp/go 
-bk restore  go2 /tmp/go 
-(cd ~/go/src; for i in **/*(^/); do cmp $i /tmp/go/$i; done)
+for b in test `bk list | grep test@ | awk '{print $1}'`; do
+    rm -rf $RESTORE_DIR
+    echo Restoring backup $b
+    bk restore $b $RESTORE_DIR 
+    (cd $SRC_DIR; for i in **/*(^/); do cmp $i $RESTORE_DIR/$i; done)
+done
+rm -rf $RESTORE_DIR
 
-rm -rf /tmp/go 
-bk restore  go3 /tmp/go 
-(cd ~/go/src; for i in **/*(^/); do cmp $i /tmp/go/$i; done)
 
 #encrypted
 export BK_PASSPHRASE=foobar
+
 rm -rf $BK_DIR
 mkdir $BK_DIR
-bk init --encrypt  
-bk backup  go ~/go/src 
-bk backup  go2 ~/go/src 
-bk backup --base go2  go3 ~/go/src
-bk list  
-bk fsck  
 
-rm -rf /tmp/go 
-bk restore  go /tmp/go 
-(cd ~/go/src; for i in **/*(^/); do cmp $i /tmp/go/$i; done)
+$BK init --encrypt  
+$BK backup test $SRC_DIR 
+sleep 1
+$BK backup test $SRC_DIR 
+sleep 1
+$BK backup --base test test $SRC_DIR
+$BK list  
+$BK fsck  
 
-rm -rf /tmp/go 
-bk restore  go2 /tmp/go 
-(cd ~/go/src; for i in **/*(^/); do cmp $i /tmp/go/$i; done)
+for b in test `bk list | grep test@ | awk '{print $1}'`; do
+    rm -rf $RESTORE_DIR
+    echo Restoring backup $b
+    bk restore $b $RESTORE_DIR 
+    (cd $SRC_DIR; for i in **/*(^/); do cmp $i $RESTORE_DIR/$i; done)
+done
 
-rm -rf /tmp/go 
-bk restore  go3 /tmp/go 
-(cd ~/go/src; for i in **/*(^/); do cmp $i /tmp/go/$i; done)
+rm -rf $RESTORE_DIR
+rm -rf $BK_DIR
